refactor(sidecar): split gateway daemon modules
